Creating LoginScreen style file
