+++ src/screens/LoginScreen.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import {
-  StyleSheet,
   Text,
   TouchableOpacity,
   KeyboardAvoidingView,
@@ -15,70 +14,7 @@ import EmailField from '../components/EmailField';
 import PasswordField from '../components/PasswordField';
 import ButtonWithActivityIndicator from '../components/ButtonWithActivityIndicator';
 import { backHandlerPop } from '../NavigationFunctions';
-
-const styles = StyleSheet.create({
-  principal: {
-    flex: 1,
-  },
-
-  content: {
-    flex: 6,
-    paddingHorizontal: 18,
-    backgroundColor: 'white',
-    justifyContent: 'center',
-  },
-
-  footer: {
-    flex: 0.5,
-    borderTopColor: '#a9a9a9',
-    borderTopWidth: 1,
-    backgroundColor: 'white',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-
-  icon: {
-    margin: 0,
-  },
-
-  Inputemail: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    borderColor: 'gray',
-    backgroundColor: '#FAFAFA',
-    borderWidth: 1,
-    borderRadius: 7,
-  },
-
-  InputPassword: {
-    marginTop: 30,
-    flexDirection: 'row',
-    alignItems: 'center',
-    borderColor: 'gray',
-    backgroundColor: '#FAFAFA',
-    borderWidth: 1,
-    borderRadius: 7,
-  },
-
-  buttonLogin: {
-    paddingVertical: 18,
-    marginTop: 50,
-    borderRadius: 8,
-    borderWidth: 1,
-    alignItems: 'center',
-    backgroundColor: '#FF9500',
-  },
-
-  buttonText: {
-    color: 'white',
-    fontSize: 20,
-  },
-
-  loading: {
-    marginTop: 50,
-    paddingVertical: 13,
-  },
-});
+import styles from '../Styles/LoginScreenStyle';
 
 export default class LoginScreen extends React.Component {
   constructor(props) {
